extract OmnipoolHooksAdapter
